Make a tweak to the ruby AST
Remove the special-purpose `FooAsgn` nodes, and replace them with a
generic `Assign` node.

In addition to being simpler (this is a net code deletion), this lets
us store better position information, since the left-hand-sides of
those assignments now become their own nodes that accurately record
the extent of *just* the left-hand-side. Previously, we were imputing
the location of the entire assignment to the left-hand-side.