Add a few annotations to the std lib
Dir.mktmpdir comes from 'tmpdir' but I chose to add it to dir.rbi
because requiring the file monkey-patches Dir, and it comes from the
stdlib. This could cause confusion if someone looked at the shim files.

On the other hand I'm not sure that Sorbet running in LSP would be aware
of the different required files, and whether or not it could correctly
inform the editors.