TreeMap over OptionalArgs too (#1389)
TreeMap over OptionalArgs too (#1389)

* TreeMap over OptionalArgs too

Before, we would just drop the default_ expressions on the floor and
replace the entire arg with a Local. Now that we're doing something with
the default_, we need to make sure it gets named like anything else.

We're only descending into OptionalArgs, because those are the only type
of arg with a nested expression (default_). Correctness then depends on
OptionalArgs being at the top level (because we skip descending into any
other arg type), so we ENFORCE this invariant.

We also need to change namer: we were moving the `method->args` into
enclosingArgsStack in preTransformMethodDef and moving it back in
postTransformMethodDef, so the treemap never saw any args to walk into.

The only reason why we were stealing the args was to sometimes append a
newly created BlockArg to the list of arguments. Instead of stealing the
entire list to append one thing, we just record the thing to append, and
append it in postTransformMethodDef.

* Enforce that we don't do double substitutions

In Substitute.cc, we were `TreeMap::apply`ing the substitution over the
`default_` of an optional arg. We don't need to do this anymore, because
the TreeMap already does it. This manifested as a crash, so we add an
ENFORCE against it happening again.

It's actually okay in a couple of cases for the 'from' and the 'to' to
be the same, because we create and perform a substitution on the fly (so
the 'from' GlobalState ID in the NameRef will be created with the 'to'
being the current GlobalState). In these cases, we get around the issue
with an optional flag to `substitute` that skips the check.

(Squashed by Merge Bot)