Elaborate optional args to T.let in resolver (#1388)
Elaborate optional args to T.let in resolver (#1388)

* Add some test files

* rbi: _ -> T.unsafe(nil)

We discussed whether we'd like to special case "_" syntax in our rbi
files, or whether we'd like to do something like this.

Most people at Stripe won't have to deal with this syntax, because we're
the ones that deal with rbi's. But if people outside of Stripe don't
like this syntax, we can always define something like `T._` in Helpers
which has type `T.untyped`.

* Update some snapshots that are different now

We used to replace args with Local's regardless of what kind of arg it
was. Now, we only replace UnresolvedIdents with Locals, which might be
nested inside some other Reference (like RestArg or OptionalArg), which
is why this test changed.

* Elaborate optional args to T.let in resolver

This is the meat of the solution. Instead of checking default args
directly, we elaborate a default arg like `exp` with type `A` to
`T.let(exp, A)` at the top of the method body. Then we match up the
loc's to get nice errors that appear to come from the right place.

* Make arg2Symbol recursive

Before, we would replace the entire argument tree with a Local.
That means we throw away the OptionalArg before it even gets to
resolver, where we can inject the T.let's into the body.

What we do now is overwrite the nested UnresolvedIdent with a Local,
preserving the argument's original structure. To do this, we needed to
take a `unique_ptr<Reference>&`, and to do that we needed to replace the
`while` loop with a recursive function.

* Extract the nested Local in cfg

Now that we're preserving the arg structure in the namer, we needed to
teach cfg how to walk into an argument and extract the Local.

* nits (comments and variable names)

* Remove the macro

* Bump pay-server SHA

* Fix exp output

(Squashed by Merge Bot)