Make Range.new generic in element type (#1441)
Make Range.new generic in element type (#1441)

* Assign ranges to temp in tests

No changes made yet; this is so we can see the diff when we actually do
make changes.

* Make Range.new generic in element type

Uses type_parameters to keep the type information around when creating
a new Range.

* Add test using #first and #last

Note in the typed-source: e and f are Integer and String, respectively.

* Range.new takes an optional param (exlcude_end)

* Bump pay-server-sha

(Squashed by Merge Bot)