Remove+merge some arguments from dispatchCall (#1455)
Remove+merge some arguments from dispatchCall (#1455)

* Collapse three dispatchCall args into one struct.

* Refactor dispatchCallWithTargs.

Make it a free function. This was never *really* a method on
ClassType; It's a helper to share logic between ClassType and
AppliedType, and AppliedType made a weird fake ClassType in order to
call into it. Instead, pass in `(this, symbol, targs)` explicitly and
make this property of it as a shared helper more explicit.

* Refactor lastAndComponent.

Repurpose `selfRef` to have the semantics of `lastAndComponent`, and
use `this` in cases where we actually meant `selfRef`. The primary
users of `selfRef` were in intrinsic methods, which actually wanted
`this` but didn't have access to it, so we add that in only for those.

I find this presentation more understandable, and also more accurate
-- `lastAndComponent` differs from `this`/the old `selfRef` in other
ways than around `OrType`/`AndType`, so the name was a
simplification (in particular, for proxy types, `lastAndComponent`/
the new `selfType` still carries the proxy)

* Add a comment.

* Pack most dispatchCall arguments into a single struct.

* selfRef -> selfType

* Move thisType to the end of the argument list

This keeps Context first and is more parallel to how I did
dispatchSymbol

(Squashed by Merge Bot)