Fix dropSubtypesOf by manually calling OrType::make_shared (#1464)
Fix dropSubtypesOf by manually calling OrType::make_shared (#1464)

We found a bug in dropSubtypesOf where dropping NilClass from an OrType
would cause the resulting type to not be a subtype of what we started
with. In particular, the bug was that

    dropSubtypesOf(NilClass | Array[Integer] | Array[String],  NilClass)

became

    Array[Integer | String]

and this is not a subtype of the input.

This unblocks merging into sorbet master.

(Squashed by Merge Bot)