Fix some showRaw() output (#1468)
Fix some showRaw() output (#1468)

* Use nodeName() everywhere

Using nodeName() everywhere make it slightly less likely that what we
print gets out of date. In particular, we were printing `Ident` when we
meant to be printing `Field` (see the test that changed)

* Add comment

I happened to find this surprising, so I figured I'd leave a comment.

* This test will change after the nodeName() change

* Ident -> Field

This is really the only thing that happened in the nodeName() change on
the previous commit.

(Squashed by Merge Bot)