Better handle `*args` in procs (#1460)
Better handle `*args` in procs (#1460)

* Tweak how LoadArg works.

We have a Symbol for the argument when constructing a LoadArg, which
we can just hold on to and use the resultType from. Previously we were
throwing that information away just to turn around and dig through
name tables to find the same symbol.

* getCallArgumentType -> getCallArguments

Return the entire argument list, as seen by the implementation, as
a (typically) tuple type. This is the interface the only remaining
caller wants, anyways, and it's cleaner to implement this way.

* Fix rest-arg handling in blocks

* Fix and test the `&blk` desugar

* Update .exp files

* Splat T.untyped in procs, too

* Forward-port the *args fix

* Refactor an `if` into Symbols.cc

* Take a `Context`

* Add a requested test

* Bump pay-server

* Bump pay-server again

(Squashed by Merge Bot)