Make sig suggestion work with private methods (#1544)
Make sig suggestion work with private methods (#1544)

* [no change] zero-with replacementLoc

This shouldn't change the meaning, but makes it easier to do future
changes (because we don't have to worry about tacking the 'def' back on)

* tools/scripts/update_exp_files.sh

Now that we're doing zero-size replacementLoc, the tests don't have the
`def` in the replacement anymore.

* [no change] Renaming some things

These names make more sense to me

* Find first character of line for autocorrect

* Add test for private / protected

* Use 'Insert' when doing zero-width replacement

* tools/scripts/update_exp_files.sh

* At least show something for zero-width loc

We might want to show something else to indicate that this loc
represents specifically the cursor being between two characters, but idk
what that character would be. `|` didn't look that good.

* tools/scripts/update_exp_files.sh

(Squashed by Merge Bot)