Don't suggest if has block arg (#1546)
Don't suggest if has block arg (#1546)

* [no change] Don't check for `def`

Now that we're not replacing the `def` (and instead using a zero-width
`replacementLoc`), we don't need to check that the line has it.

* Add test for suggesting sig with block

Note how the suggested sig has `<blk>` as the name of the block arg,
which is invalid syntax.

Also, we don't really want to show sigs for blocks right now because
we'll never guess a type for them.

* Don't print if has block arg

We don't want to suggest a sig when the method has a block argument:

- Sometimes we synthesize a block arg, and name it `<blk>` which is not
  a valid identifier name.
- We're not attempting to infer types for them, and we don't want people
  to think that procs can't be typed.

* tools/scripts/update_exp_files.sh

This change verifies that we are no longer suggesting when the
suggestion would have a block argument.

* any_of -> absl::c_any_of

(Squashed by Merge Bot)