Don't infinitely loop in guessArgumentTypes (#1576)
Don't infinitely loop in guessArgumentTypes (#1576)

* Don't infinitely loop in guessArgumentTypes

On the test case provided, we were seeing an infinite loop because glb
did not return referencially equal types if it was possible to reuse
either t1 or t2 as the greatest lower bound of the two.

guessArgumentTypes uses this information to figure out when it has
converged to a fixed point.

The alternative here (if we don't want to commit to this invariant) is
to use structural equality rather than referencial equality on the
types in guessArgumentTypes. cc @dmitry for why we didn't go for this.

* Don't use so many bool's

(Squashed by Merge Bot)