Change what we capture for the sig suggestion tests (#1674)
Change what we capture for the sig suggestion tests (#1674)

* Only capture the file changes

In particular, for these two test files, don't capture the error output.
We already have tests for the error output.

We also don't need the extra song and dance we previously needed for
these tests to make sure we the filename in the error output was stable
(as oppose to whatever temp filename we gave).

* scripts/tools/update_exp_files.sh

Note how this is strictly a deletion of captured output.
The part below the `---` line remains intact and unchanged.

(Squashed by Merge Bot)