More flags (#1671)
More flags (#1671)

* int -> u4

We want to use u8 instead of int because we're doing bitwise operations
and want to know specifically how many bits of wiggle room we have. The
number of bits in an int is platform specific.

* Re-index the Symbol flag bits

We're packing three different kinds of flags into separate ranges with
the u4's below:

    0x0000'0000
      ├▶    ◀┤└─ Applies to all types of symbol
      │      │
      │      └─ For our current symbol type, what flags does it have?
      │         (New flags grow up towards MSB)
      │
      └─ What type of symbol is this?
         (New flags grow down towards LSB)

* Expand all constants to 32 bits

(Squashed by Merge Bot)