Correctly model override / overridable suggestions (#1673)
Correctly model override / overridable suggestions (#1673)

* parentNeedsOverridable & childNeedsOverride

This is the core change.
It uses the table here[1] (compiled from auditing pay-server).
We don't need override in most suggestions. Not putting them makes it
easier to propagate sigs everywhere.

* Add implementation to trigger overridable suggestion

This is the only case in which we trigger the suggestion to add
overridable.

* tools/scripts/update_exp_files.sh

* tools/scripts/update_exp_files.sh

(Squashed by Merge Bot)