`floor` and `ceil` take a precision. (#1702)
`floor` and `ceil` take a precision. (#1702)

If called with no arguments, they always return `Integer`. If called
with one argument, they tend to return either `Integer` or
`T.self_class` depending on whether or not the argument is zero. Just
call this `Numeric`, which works fine in most cases.

(Squashed by Merge Bot)