Standardize usage of std:: namespace (#1752)
Standardize usage of std:: namespace (#1752)

* Remove 'using namespace std' from header files

We don't like to include this in headers because it pollutes everywhere
the file is included, transitively, which can sometimes cause
unsuspecting name collision errors.

`using namespace std` is fine in `*.cc` files though.

* gsed -i -e 's/std:://' **/*.cc

* git checkout master third_party

* git checkout master core/tools/generate_names.cc parser/tools/generate_ast.cc payload/text/tools/generate_payload.cc

These files write `std::` into the files they generate, which we want to
preserve.

* git diff-review | diff-locs --all | grep '^-' | cut -c '2-' | multi-sed - 's/move/std::move/g'

We must fully-qualify std::move because it's ambiguous with absl::move
when you remove it, even though the two names are the same function.

* manual: Add 'using namespace std;' to some files

* git diff-review | diff-locs | multi-sed - 's/^template class /template class std::/'

We need to fully-qualify all types in template declarations

* manual: Fix some template declarations that didn't match our regex

All of these have std:: nested within a type application, and I didn't
feel like writing a sed which would fix them.

* manual: Don't use min / max for local variable names

* git checkout master parser/Builder.cc

This file was giving me a lot of trouble, because it's very different
from the rest of our files.

* manual: Rename istream and ostream members

* manual: For some reason, these had to be fully qualified

* tools/scripts/format_cxx.sh

(Squashed by Merge Bot)