Make define_method and define_singleton_method the same (#1785)
Make define_method and define_singleton_method the same (#1785)

* Make define_method and define_singleton_method the same

I think that blk was typed as Proc before was the reason why we were
seeing a bunch of weird errors in pay-server. I don't want to force
everyone else to have to go through this when trying to use
define_singleton_method in the future.

* tools/scripts/update_exp_files.sh

(Squashed by Merge Bot)