Refactor QueryResponse (#1803)
Refactor QueryResponse (#1803)

Refactor QueryResponse. It is now a wrapper around a std::variant type with helpers for extracting common properties and for determining which variant it is (as std::visit and friends aren't available on a Mac).

I also moved Query, QueryResponse, and friends from sorbet::core into sorbet::core::lsp.

Prior to this refactoring, QueryResponse was a poor man's union type. It contained a variety of fields, with only a subset in use depending on the QueryResponse's kind. It was hard to construct (as you had to pass in dummy values for inapplicable fields), and hard to code against as you had to remember which fields were active for each kind of response.

(Squashed by Merge Bot)