Fix order dependence error with type aliases (#1867)
Fix order dependence error with type aliases (#1867)

* Fix order dependence error with type aliases

By having `move(...)` below, we were implicitly changing anything that
had a reference to this old pointer. In particular, the `todo_` list
could still have a reference to the old `nesting_`. Consdier this
example:

    module Private
      StringAlias
    end

    StringAlias = T.type_alias(String)

In the first walk, the StringAlias on line 2 will not be resolved
(because it isStaticTypeAlias) and will get added to the `todo_` list.
When added to the `todo_` list, its job will have a `scope` that points
to this `nesting_` object.

Then in the postTransformClassDef, we'd do this

    nesting_ = move(nesting_->parent)

trick, which causes todo[0].job->scope->parent to become `nullptr`
(because it's the same as `nesting_->parent`). So when we dequeue the
`todo_` job, the `nesting_` scope has changed!

We guarantee agains this happening in the future by marking the
`parent_` and `scope` fields `const`.

* Add failing test

* tools/scripts/update_exp_files.sh

I'm not quite sure why this changed, but judging by the output, it seems
correct. Basically we reported the missing constant `Foo` three times
before, and one time now. And note that even in the "Did you mean"
section, we pointed at the right `Foo`.

(Squashed by Merge Bot)