Make computeMinMaxLoops more explicit (#1908)
Make computeMinMaxLoops more explicit (#1908)

* Make computeMinMaxLoops more explicit

In one of my recent changes I was grepping around for where minLoops was
being populated and though my grep found this in the results, I
discounted it because it didn't look like something of the form
`minLoops[x] = y`.

In fact, this **was** the place where `minLoops` was being populated,
and the way it worked is through C++ references, which I found
non-intuitive. Rewriting this to use normal local variables + operator[]
for clarity.

Also dmitry seemed to think that this might even be faster, because the
C++ compiler wouldn't have to reason about invalidating refs and could
instead reason 100% locally about temps.

* PR feedback

(Squashed by Merge Bot)