Update autogen format to make pay-server faster. (#1913)
Update autogen format to make pay-server faster. (#1913)

* Stop using msgpack extensions.

Gate the behavior behind a `--autogen-version` flag. Because of our
two-sided CI, it's not actually currently possible to directly change
the autogen format, since this branch has to pass on an old pay-server
branch before we can merge it to a new pay-server branch.

Adding versioning lets us support two formats at once, and we'll add
`--autogen-version=1` to pay-server when we update it to support the
new format.

* Remove some defunct fields from autogen output.

These were constant and left behind for compatibility with
old-autogen. Strip them out in version 1.

* update_exp_files

(Squashed by Merge Bot)