Synthesize block args in desugar not namer (#1909)
Synthesize block args in desugar not namer (#1909)

* Synthesize block arg when we see block_given? / yield

Remove block arg creation from namer, and ENFORCE

Remove TODO

Use Loc::none() for synthesized block arg

Always synthesize a block arg

Silence 'Type not specified' error for <blk>

* Change sig suggestion to be aware of <blk>

* Expectation changed

* Change override check know every method has a blk

* Populate blockPreType before dispatching to intrinsic

This has to be here because Class_new is an instrinsic, and it
dispatches to the initialize method on the child class. In the other
order, blockPreType would be overwritten by looking up the block arg of
the outer Class.new method definition, which is untyped because it
wasn't mentioned in the sig.

Specifically, this change needed to be here because without it, it broke
the T.assert_type!(...) call in `test/testdata/infer/block_arg.rb`:

    ConstructorBlock.new do |sym|
      T.assert_type!(sym, Symbol)
      sym.to_s
    end

* [part 1] tools/scripts/update_exp_files.sh

In namer we used to fill in the type of the created symbol when we
created a discovered block arg. Now, we don't, because it's just like
any other param in a method def without a sig.

* [part 2] tools/scripts/update_exp_files.sh

* [part 3] tools/scripts/update_exp_files.sh

We expect to see an extra load_arg instruction, because we added an arg.

* [part 4] tools/scripts/update_exp_files.sh

These updates are to LSP hover. For the time being, we'll be showing
people that every method has an extra `<blk>: T.untyped` argument, which
is not a great state to be in. Dmitry and I IRL'd and think that we
should handle this in a follow up PR later, closer to when we ship LSP
hover (we're shipping diagnostics first).

* [part 5] tools/scripts/update_exp_files.sh

We relaxed the constraint in SigSuggestion to be able to suggest sigs
that have block args (we use to unilaterally reject suggesting sigs for
methods with block args). Now, instead of rejecting them, we'll actually
just hide the synthesized `<blk>` paramter.

* [part 6] tools/scripts/update_exp_files.sh

The change to add a block arg with a none() Loc revealed that we had a
bug in our locs for reporting errors in methods with no body. When we
fixed that bug, this snapshot changed (the Return instruction's loc is
now correctly "the entire method" not "the loc of the last argument in
the method").

* Make namer idempotent

Namer replaces args with locals, so to make namer idempotent,
we need to be able to handle Locals here.

* Expectation changed for namer unit test

Now that we're creating a block,

- we expect the argument size to be 1, not 0
- we expect the number of symbols entered to increase by 2, not 1,
  because now there's a new block arg for the file-level static init.

* Move yield / block_given? to desugar

I decided to do this because the namer idempotency tests were failing.
After namer runs, we convert all arguments into locals, which means that
we can't check that the back arg is a BlockArg (it's a Local) before
pushing it onto the blockArgStack.

I could have hacked around this, but honestly I'd been getting tired of
all the hacks in this PR, so I decided to rip out blockArgStack
entirely. If there's no blockArgStack, then there's no need to populate
it, and no need to enforce that we populate it correctly.

Instead, I just moved yield / block_given? to be desugared in Desugar.cc

* PR feedback: hide block arg from redefinition counts

* Fix bug from pay-server

pay-server revealed that there was a bug in our implementation for
override checking. The parent had `&default`, the child had `&_default`,
and we were checking whether the argumentName() was different and the
parent was not `<blk>`. This was too restrictive, so I relaxed it.

cc @dmitry

* Remove redundant code from ast/Helpers.cc

* Bump pay-server

* Re-record test expectations after rebase

(Squashed by Merge Bot)