Type assignment expressions the way Ruby does
The return value of an assignment expression is the value
that is on the right hand side.

Before this commit, we were using the return type of the
callee as the return type of the assignment expression.

See: https://github.com/ruby/spec/blob/53daec9f164e33d291d03fa2c23440568c2b2657/language/send_spec.rb#L217