Remove useless selfSym from CFG (#2008)
Remove useless selfSym from CFG (#2008)

* Remove useless selfSym from CFG

We used to need this here because it's location was guaranteed to be
non-none, and sometimes a Loc::none() as the last instruction in a block
caused us to misprint some errors. (#1909)

Even before that, we used to need this here because the `load_arg`
instruction referenced this temp. We removed that when we refactored how
undeclared variables are reported (#1917).

While these two things were fixed previously, we kept the symbol in to
avoid having the snapshots all change. The only thing this change does
it to removed the useless variable and update the snapshots (most
changes are just temps getting renumbered).

In fact, we don't even see the `self` instruction itself being removed
from the snapshots, because it was already removed by removeDeadAssigns
pass.

* tools/scripts/update_exp_files.sh

(Squashed by Merge Bot)