Add ENFORCE that every symbol has a block arg (#2000)
Add ENFORCE that every symbol has a block arg (#2000)

* Unconditionally enterMethodArgumentSymbol for blk

* Conditionally add the block arg

These are places where enterMethodSymbol might return an existing
symbol, so we need to check if something was created for the first time
and only add the block argument then.

* Fix some comments in calls.cc

* Fix method overloads with blocks

We used to erase the block argument from a method symbol's overloads
when that specific overload didn't mention it. This was significant for
overload resolution, because the presence of a block at the call site
was checked against whether the overload sig mentioned a block argument.

Now that every symbol has a block argument, we instead check for whether
the block argument is synthetic (i.e., created by us), and assuming that
if we didn't create it, it must have been mentioned in the sig.

* Fix how we enter intrinsics

This is the cause of the bug we saw on pay-server master, which we fixed
by deleting code:

<https://git.corp.stripe.com/stripe-internal/pay-server/pull/132001>

Here, we fix the bug, and add back a test to make sure that we don't
regress.

* tools/scripts/update_exp_files.sh

* Add new ENFORCE for method argument symbols

* Mark LSPSlowPathTest as manual

* tools/scripts/format_build_files.sh

* fixup

(Squashed by Merge Bot)