Make optional: false the default in Sorbet (#2024)
Make optional: false the default in Sorbet (#2024)

* Remove isNilable

`optional: false` is the default and the only way to specify something
is nilable is to write `T.nilable`.

* Record behavior changes

I verified these with @jerry to make sure that these changes correspond
to the actual changes that happened in the runtime system. Basically,
having `optional: false` is ignore because it's the default, and not
using `T.nilable(...)` in the prop type means that you're not nilable.
Basically, we never need to call `mkNilble` anymore.

* tools/scripts/update_exp_files.sh

This is a pretty random test; I'm not really sure what it's trying to
test, but looking at `same-loc.rb`, I'm pretty sure that these errors
shouldn't be reported anymore, so I'm 👌 to delurking these.

* Change error: expectations

* tools/scripts/update_exp_files.sh

* Add tests for ifunset

* tools/scripts/update_exp_files.sh

* Foreign relationship is nilable

* Add test for T.nilable in the prop type

* tools/scripts/update_exp_files.sh

* Bump pay-server SHA

* tools/scripts/update_exp_files.sh

(Squashed by Merge Bot)