Remove support for `optional` in prop definitions. (#2046)
Remove support for `optional` in prop definitions. (#2046)

* Remove isOptional

The storage team has removed optional as a prop method.
Everything must be `prop` with `T.nilable`.

* Remove usage of optional in tests

* tools/scripts/update_exp_files.sh

* Clean up prop.rb test (#2047)

* Sorbet does not look at the 'optional:' key anymore

* scripts/tools/update_exp_files.sh

Note that the only thing that changed here is the end pos of the loc.

* Remove some redundant props

The only thing that used to distinguish these tests was the `optional:`
bits. Now that we're not looking at that, we can just remove them.

* scripts/tools/update_exp_files.sh

The only changes here are to:

- the trees (ofc the trees change, because we deleted a bunch of stuff)
- the **locs** and the **names** in the symbol table

Imporantly, none of the types in the symbol-table exp file changed.

* Convert everything to T.reveal_type

I noticed that there were a lot of useless tests using `T.assert_type!`
in this file. There were a lot of places where the original author
likely meant to assert "this is not-non-nil" and the way they did that
was by asserting that it was both nilable and that there was an error if
trying to use it as non-nil. :psyduck:

Instead, for these tests it's more valuable to know when something's
type has changed at all, not whether it's still a subtype of some
mentioned type. `T.reveal_type` works better for this.

This change is just a dumb substitution. See the next commit for the
actual clean up that I wanted to make.

* Get rid of redundant tests

We can also remove most of the commented out tests, because we can
actually check them with T.reveal_type.

* Fix expectation

T.reveal_type converts "Array" into "T::Array[T.untyped]" when printing.

* Fix array: expectation

Looks like we're doing something wrong for the array: rule. I'm not sure
what the expected behavior here is, but I don't really want to fix it.

At least it's more explicit than it used to be! So the next person who
comes along can understand the current state better.

* Looks like enum: is explicitly untyped

* scripts/tools/update_exp_files.sh

Again, only the trees are changing here. The symbol-table remains
unchanged.

(Squashed by Merge Bot)