Fix bug: Prevent duplicate responses to the same request when an error occurs in setupLSPQueryByLoc (#2066)
Fix bug: Prevent duplicate responses to the same request when an error occurs in setupLSPQueryByLoc (#2066)

* Fix bug: Prevent duplicate responses to the same request when an error occurs in setupLSPQueryByLoc.

Previously, if setupLSPQueryByLoc sent an error, callers would proceed as if they received an empty result to their query.

definition_error.rec did not catch this bug; I assume it ignores any messages that occur after it has checked all of the messages in the recording?

Now, callers receive either a TypecheckRun or a pair of (ResponseError, GlobalState). The former indicates a successful run, the latter indicates an issue.

(Squashed by Merge Bot)