Show unique number in symbol-table output (#2083)
Show unique number in symbol-table output (#2083)

* Show unique number in symbol-table output

We already show it in the symbol-table-raw output, and I'd like it in
the symbol-table output, but I don't want it to be for every unique name
kind (this would mean that every time you added a `do; end` block in a
file, you'd have to update the exp files.)

* Use custom UniqueNameKind for mangleRenameSymbol

* tools/scripts/update_exp_files.sh

* Fix two direct comparisons to Namer

There were two places where we were directly comparing to a specific
enum (so we don't get exhaustiveness checking). Both these places were
looking to see if the name was actually some sort of class name.
Since we're only using `Namer` for methods now (`<block>` and
`<static-init>`), let's replace with `... == MangleRename` instead of
`... == Namer || ... == MangleRename`.

(Squashed by Merge Bot)