Fix mixes_in_class_methods in --stress-incremental-resolver (#2101)
Fix mixes_in_class_methods in --stress-incremental-resolver (#2101)

* Fix mixes_in_class_methods in --stress-incremental-resolver

We store the mixes_in_class_methods class as a member named
`<class methods>`, so if you ever declared `mixes_in_class_methods`
twice, we'd effectively overwrite the previous one, and so we emit this
error. However, it's not really overwriting if the thing we're about to
write into the `<class methods>` member is what already exists there.

* Add a test for mixes_in_class_methods_twice

* Update existing test

* Make error message better

(Squashed by Merge Bot)