Stub the toMethod of alias_method if doesn't exist (#2128)
Stub the toMethod of alias_method if doesn't exist (#2128)

* Stub the toMethod of alias_method if doesn't exist

This lets us generate correct rbi's. Take for example this code:

    define_method(:transaction_currency) do; end
    alias_method :currency, :transaction_currency

Sorbet won't see #transaction_currency, and as a result, won't create
the method alias. When we regenerate missing methods, we'll write out

    def transaction_currency; end
    def currency; end

Then the error becomes: we're redefining currency as a method alias.

To get around this, instead of not entering a method symbol for
currency at all, we enter a method symbol, but alias it to a stub method
that we entered in GlobalState. This way the missing method script will
see that Sorbet has a symbol for `currency` and only serialize a method
for transaction_currency in the rbi.

* Make the stub permissive

* Make methods for all constants

(Squashed by Merge Bot)