autogen: Pack ranges using 64-bit integers. (#2133)
autogen: Pack ranges using 64-bit integers. (#2133)

* autogen: Pack ranges using 64-bit integers.

Some files have lengths larger than 2**16 bytes, so the old version
would wrap.

I'd used 32-bit ints previously because I wanted to avoid Ruby going
to bigints and allocating on the heap, but in practice that will only
happen if the actual *values* are above 2**62 or so, so this should be
fine.

* Add a cast

(Squashed by Merge Bot)