Remove typeAliases map (#2152)
Remove typeAliases map (#2152)

* Boilerplate: make typeAlias into a bool

This PR is kind of big, so I've broken it up into four main commits and
then a bunch of stragglers. The four steps are basically:

1. Convert ConstantLit::typeAlias from storing an Expression to a bool
2. Delete resolver's typeAliases map
3. Store type alias information using the typeAlias boolean + the type
   alias symbol's resultType.
4. Update the downstream users of type alias information now that we're
   storing a type, not a tree for each resolved type alias.

So this PR is just boilerplate: changing the type of typeAlias from
ast::Expression to bool.

* Boilerplate: delete typeAliases map

These changes look kind of ugly because of clang-format, but it's really
just deleting the typeAliases map and updating every function signature
and call.

* Use the new resultType in CFG + Infer

Instead of type aliases storing resolved trees, type aliases store
parsed types, so we can create an Alias instruction and treat it
specially in environment.cc: if the symbol of this alias is a type
alias, the result type of this whole instruction is a MetaType wrapping
the type alias's type.

This kind of magically works, and I didn't push too deep into why it
works or even attempt to rationalize it from first principles (i.e.,
what does it mean for this instruction to be a MetaType?).

We also use the resultType when parsing types in type_syntax.

* update_exp_files: nullptr -> false

typeAlias is a bool now, so the snapshots all change.

* Write the resolver.cc header comment I wish I had

It was out of date (it referenced the non-existing type aliases map) and
after I started reading into it I realized that most of it was out dated
and or misleading, so I just rewrote it.

* getResultType: unique_ptr<Expression> -> Expression&

When I added the TypeAliasResolutionItem job type, I had a pointer, not
a unique_ptr, because the pointer was owned by an actual tree.

Rather than make a new unique_ptr, hand it to getResultType, and then
"unsafely" release ownership after the call to getResultType, I just
opted to make getResultType not take a pointer.

* Populate the typeAlias boolean & resultType type

We add a new kind of resolution item + todo list to resolve type alises.
To resolve a type alias like

    A = T.type_alias(Integer)

we need the symbol for `A` and a reference to the tree inside the
type_alias method call.

A type alias can be resolved if all the constant symbols in the tree are
resolved (see `isFullyResolved`). If it can, we give the tree to
`TypeSyntax::getResultType` and store the resulting type directly onto
the symbol for `A`.

The invariant for whether a ConstantLit is done being resolved is now a
bit tricky; either:

- it's a type alias, and the symbol's resultType is non-null, or
- it's a normal constant, and the symbol merely exists.

* Change error message

It's possible now that "expands to an infinite type" isn't the only way
to trigger the unresolved type alias error now. Specifically:

    A = T.type_alias(B)

^ this code will also trigger it, because `A` is a type alias, but we
haven't set it's resultType, because we haven't resolved `B` to a
constant. (There is already a test for this behavior elsewhere; this
change just updates the existing error messages).

* New error: Unable to resolve type alias

This was hinted at in the last commit. We emit a new error for
HydraResource being an unresolved type alias.

* Update autogen snapshot

We used to have two references to the DOES_NOT_EXIST: the one in the
tree, and the one we dup'd onto the typeAlias field (remember it used to
be a tree before this PR changed it to a bool). That tricked autogen
into serializing out two references to it. Now that the dup'd tree is
gone, one of the dup'd entries goes away too. (Expand the diff lines
right above this hunk to see).

* T.untyped -> typed

Lo and behold, it worked! These snapshots confirm that we're seeing the
resultType on type-alias symbols as being set to the type alias's type.

* Add a test for IOLike (#2166)

This is is a type alias in the payload; we used to not handle these
correctly, but after the aliases refactor, we do.

(Squashed by Merge Bot)