Weaken ENFORCE in isTypeAlias (#2187)
Weaken ENFORCE in isTypeAlias (#2187)

* {is,set}StaticTypeAlias -> {is,Set}TypeAlias

* Weaken ENFORCE in isTypeAlias

I'd like to be able to ask whether anything that looks like a constant
is a type alias. It made the code pretty verbose to always check
`isStaticField() && isTypeAlias()`. There are three constant-like symbol
kinds, and I think that it's reasonable to ask whether any of them is a
type alias (though it should still only be possible to set a static
field to be a type alias).

* Move comment inside method

(Squashed by Merge Bot)