Integrate Watchman with Sorbet. (#2028)
Integrate Watchman with Sorbet. (#2028)

Integrates Watchman with Sorbet to watch the file system for file updates. Before this change, Sorbet was blind to changes happening on the file system after it started up.

### CLI Changes

When Sorbet is run with `--lsp`, Sorbet starts a Watchman thread that invokes the Watchman executable and subscribes to updates in:

- The current working directory if no directory is passed to Sorbet, or...
- A single directory passed to Sorbet.

Sorbet and Watchman communicate via Watchman's JSON protocol.

Watchman can be disabled with `--disable-watchman`, so it is still possible to run LSP without Watchman.

### Error Conditions

If Watchman is enabled, Sorbet will exit with a non-zero exit code in the following situations:

* If Sorbet is passed multiple files or directories. The Watchman implementation currently assumes a single directory
* If Watchman fails to start, or if the Watchman process exits unexpectedly.

In both cases, Sorbet advertises the `--disable-watchman` option.

### Other Changes

* `sorbet --lsp` requires a single folder input.
* Contains LSP protocol-level tests for Watchman integration.
* JSON serialization/deserialization logic can now expose JSON fields under different C++ names, making it possible to convert snake_case fields to CamelCase fields.
* Sorbet merges multiple pending Watchman updates into a single update. It performs this merging in the same loop that merges consecutive textDocumentDidChange events.
* Sorbet now properly handles input folders with trailing forward slashes.

(Squashed by Merge Bot)